remove step plugin & join GoldenMember
