+++ src/components/MemberAccount/MemberBoard/MemberCSS/MemberCoupon.css
@@ -11,7 +11,7 @@
 .MChistoryTitle {
   font-size: 32px;
   color: var(--fontWhite);
-  margin: 25px 40px 10px 40px;
+  margin: 40px;
 }
 
 .MCmemberBoard hr {
