+++ src/components/MemberAccount/MemberBoard/MemberCSS/MemberHistory.css
@@ -59,7 +59,7 @@
 .MHhistoryTitle {
   font-size: 32px;
   color: var(--fontWhite);
-  margin: 25px 40px 10px 40px;
+  margin: 40px;
 }
 
 .MHmemberBoard hr {
