+++ src/components/MemberAccount/MemberBoard/MemberCSS/MemberNotes.css
@@ -11,7 +11,7 @@
 .MNnotesTitle {
   font-size: 32px;
   color: var(--fontWhite);
-  margin: 25px 40px 10px 40px;
+  margin: 40px;
 }
 
 .MNmemberBoard hr {
