+++ src/components/MemberAccount/MemberBoard/MemberLevel/Golden.css
@@ -1,95 +1,52 @@
-:root {
-  --main-color: #1d3557;
-  --second-color: #457b9d;
-  --third-color: #abdadc;
-  --forth-color: #f1faee;
-  --fifth-color: #e63946;
-  --sixth-color: #fff579;
-  --blockLight: #c2ccc4;
-  --blueLine: #00a2e0;
-  --fontWhite: white;
-  --btnHover: #01619c;
-}
-body {
-  background: var(--forth-color);
-  width: 100%;
-  height: 100vh;
-}
-nav {
-  width: 100%;
-  height: 66px;
-  background: red;
-}
-.MAactivity-wrapper {
-  width: 100%;
-  height: 100vh;
-  overflow-y: auto;
-}
-
-.MAnavHead {
-  height: 65px;
-  background: var(--main-color);
-}
-
-.MAmemberAside {
-  max-width: 100vw;
-  margin: 0px auto 100px;
-}
-
-.MAmemberPage {
+.memberPage {
   display: flex;
   max-width: 1280px;
   margin: 0px auto;
 }
 
-.MAmemberCentreTitle {
+.memberCentreTitle {
   width: 280px;
   height: 80px;
 }
 
-.MACentreTitle {
+.CentreTitle {
   font-size: 70px;
   font-weight: bold;
   color: var(--main-color);
   margin-top: 50px;
 }
 
-.MAmemberCentreList {
+.memberCentreList {
   width: 200px;
   margin: 0px auto;
 }
 
-.MAmemberCentreList ul {
+.memberCentreList ul {
   padding-left: 0;
   flex-direction: column;
   list-style-type: none;
 }
 
-.MAmemberCentreList ul li a {
+.memberCentreList ul li a {
   font-size: 28px;
   text-decoration: none;
   display: block;
   color: var(--main-color);
   width: 100%;
   margin-top: 40px;
-  cursor: pointer;
 }
 
-.MAmemberCentreList ul li a:hover {
-  font-size: 32px !important;
-  color: var(--fifth-color) !important;
-}
-.MAactiveList {
+.memberCentreList ul li a:hover {
   font-size: 32px !important;
   color: var(--fifth-color) !important;
 }
 
-.MAactiveList a {
+.activeList a {
   font-size: 32px !important;
   color: var(--fifth-color) !important;
 }
 
-.MAmemberBoard {
+.memberBoard {
   position: relative;
   height: 750px;
   width: 900px;
@@ -99,25 +56,22 @@ nav {
   border-radius: 20px;
 }
 
-.MAmemberAcount {
+.memberAcount {
   display: inline-flex;
-  padding: 20px 0;
 }
 
-.MAacountTitle {
+.acountTitle {
   font-size: 32px;
   color: var(--fontWhite);
   margin: 20px 40px;
 }
 
-.MAacountSubhead {
+.acountSubhead {
   color: #cccccc;
-  /* line-height: 50px; */
-  height: 30px;
-  margin: auto 0;
+  line-height: 50px;
 }
 
-.MAmemberBoard hr {
+.memberBoard hr {
   width: 90%;
   height: 2px;
   border: none;
@@ -125,89 +79,90 @@ nav {
   margin-top: 0px;
 }
 
-.MAmemberAccountInformation {
+.memberAccountInformation {
   display: inline-flex;
-  padding: 0px 40px;
+  padding: 10px 40px;
   width: 90%;
+  /* height: 80%; */
 }
 
-.MAaccountInformation {
+.accountInformation {
   width: 70%;
   height: 100%;
 }
 
-.MAaccount {
+.accountInformation a {
+  text-decoration: underline;
+}
+
+.account {
   display: inline-flex;
   align-items: flex-end;
   width: 550px;
   padding-top: 30px;
 }
 
-.MAaccount a {
+.account a {
   text-decoration: none;
 }
 
-.MAlistTitle {
+.listTitle {
   width: 150px;
   color: var(--fontWhite);
   font-size: 24px;
 }
 
-.MAcontent {
+.content {
   font-size: 20px;
-  color: #cccccc;
+  color: #ccc;
 }
 
-.MAchange {
+.golden {
   font-size: 20px;
   color: var(--sixth-color);
   margin: 0px 20px;
-  white-space: nowrap;
 }
 
-.MAchange:hover {
-  text-decoration: underline;
-  color: var(--sixth-color);
-}
-
-.MAmember {
-  margin-left: 20px;
-  color: var(--fontWhite);
+.change {
   font-size: 20px;
-}
-
-.MAsmallChange {
-  margin-left: 20px;
-  font-size: 16px;
-}
-.MAitemCheck {
+  color: var(--sixth-color);
   margin: 0px 20px;
 }
 
-.MAsmallChange img {
-  width: 16px;
-  height: 16px;
-}
-
-.MAsmallChange {
+.change:hover {
   text-decoration: underline;
 }
 
-.MAdateInput {
+.dateInput {
   border-radius: 5px;
   margin: 0 10px;
 }
 
-.MAheadPicture {
-  width: 30%;
-  height: 300px;
+.headPicture {
+  /* width: 30%; */
+  min-width: 200px;
+  height: 270px;
+  margin: 20px auto;
 }
 
-.MAheadPicture1240 {
+.headPicture1240 {
   display: none;
 }
 
-.MApictureGo {
+.picture {
+  position: relative;
+  /* height: 100%; */
+  height: 200px;
+}
+
+.picture img {
+  position: absolute;
+  width: 200px;
+  height: 200px;
+  top: 0;
+}
+
+.pictureGo {
   width: 200px;
   height: 200px;
   border-radius: 50%;
@@ -216,7 +171,14 @@ nav {
   margin: auto;
 }
 
-.MApictureCard label {
+.pictureCard {
+  width: 200px;
+  height: 100%;
+  display: block;
+  flex-direction: column;
+}
+
+.pictureCard label {
   border-radius: 8px;
   background: #6d695f;
   width: 120px;
@@ -225,20 +187,20 @@ nav {
   cursor: pointer;
   font-size: 16px;
   color: var(--fontWhite);
-  margin: 10px 25%;
+  margin: 10px 40px;
   line-height: 36px;
   text-align: center;
 }
 
-.MApictureCard label:hover {
+.pictureCard label:hover {
   background: #777777;
 }
 
-.MApictureCard input[type="file"] {
+.pictureCard input[type="file"] {
   display: none;
 }
 
-.MAbtn {
+.btn {
   background: var(--second-color);
   color: var(--fontWhite);
   width: 240px;
@@ -247,24 +209,20 @@ nav {
   font-size: 24px;
 }
 
-.MAbtn:hover {
-  color: var(--fontWhite);
+.btn:hover {
   background: var(--btnHover);
 }
 
-.MAList1240 {
+.List1240 {
   display: none;
 }
 
 @media screen and (max-width: 1240px) {
-  .MAmemberAside {
-    margin: 20px auto;
-  }
-  .MAmemberCentreTitle {
+  .memberCentreTitle {
     margin: 20px auto;
   }
 
-  .MACentreTitle {
+  .CentreTitle {
     margin: 0px auto;
     font-size: 70px;
     font-weight: bold;
@@ -272,81 +230,68 @@ nav {
     width: 280px;
   }
 
-  .MAmemberAccountInformation {
+  .memberAccountInformation {
     max-width: 70vw;
   }
 
-  .MAaccountInformation {
-    width: 100%;
-  }
-  .MAmemberAcount {
-    margin-left: 20px;
-  }
-
-  .MAaccount {
+  .accountInformation {
     width: 100%;
   }
 
-  .MAheadPicture720 {
+  .account {
+    width: 100%;
+  }
+
+  .headPicture720 {
     display: none;
   }
 
-  .MAheadPicture1240 {
+  .headPicture1240 {
     display: block;
-    margin: 10px auto;
-    height: 250px;
   }
 
-  .MAmemberBoard {
+  .pictureCard {
+    margin: 0px auto;
+    padding: 20px 0 0 0;
+  }
+
+  .memberBoard {
     margin: 0px auto;
     max-width: 80vw;
     height: 100%;
   }
 
-  .MAList1240 {
+  .List1240 {
     display: block;
     width: 100%;
   }
 
-  .MAList1240 ul {
+  .List1240 ul {
     display: flex;
     flex-direction: row;
   }
 
-  .MAList1240 ul li {
+  .List1240 ul li {
     margin: 0px auto;
   }
 
-  .MAList1240 ul li a {
+  .List1240 ul li a {
     display: flex;
     flex-direction: column;
     margin: 0px auto;
   }
 
-  .MAList1240 ul li a img {
+  .List1240 ul li a img {
     width: 40px;
     height: 40px;
     margin: 0 auto;
   }
 
-  .MAmemberPage {
+  .memberPage {
     flex-direction: column;
   }
 
-  .MAcontainerList {
+  .containerList {
     display: none;
-  }
-  .MApictureCard {
-    padding: 20px 0 0 0;
-  }
-
-  .MAbtn {
-    margin: 0px auto;
-  }
-}
-
-@media screen and (max-width: 760px) {
-  .MAmemberAccountInformation {
-    padding: 10px 30px;
   }
 }
