+++ src/components/auth/board/Form/MemberJoin2.css
@@ -7,7 +7,7 @@
   width: 360px;
   height: 30px;
   color: var(--fontWhite);
-  font-size: 16px;
+  font-size: 20px;
   display: flex;
   justify-content: space-between;
   line-height: 40px;
@@ -22,8 +22,8 @@
 
 .ML2dateInput {
   background: #dcdcdc;
+  font-size: 16px;
   border-radius: 5px;
-  /* width: 27%; */
 }
 
 .ML2gender {
